=== polymake 4.0 ===
-- general --
  * Data file format switched from XML to JSON
  * Basic datatype for all sizes and indexes in C++ is now 'Int' which is 'long int'
  * Object class in C++ renamed to BigObject

-- distribution --
  * support for clang 9

-- interfaces --
  * libnormaliz:
    - Update bundled version to 3.8.3
    - configure: check for extra libraries
  * cdd: Fix configure.pl for fedora installed cdd
  * porta: disable convex hull rules, can still be called via porta_primal and porta_dual
  * TOSimplex: various improvements

-- common --
  * PolyDB: new layout of collections and metadata, no XML or perl code stored with data anymore
  * PolyDB: using JSON schemas for collection definition
  * PolyDB: new object-oriented user interface

-- fan --
  * HyperplaneArrangement: Refactor
    - Remove property INPUT_HYPERPLANES
    - Property HYPERPLANES allows redundant hyperplanes now
    - Introduce simple version of facet_arrangement

-- fulton --
  * ToricDivisor: Introduce property VERY_AMPLE
  * New standard construction for weighted projective space

-- polytope --
  * Polytope: new property FACET_VOLUMES
  * beneath_beyond: fixed bug in corner-case for non-fulldimensional input
  * minkowski_sum_fukuda: improvements and cleanup

+++ lib/core/include/color.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 1997-2019
+/* Copyright (c) 1997-2020
    Ewgenij Gawrilow, Michael Joswig, and the polymake team
    Technische Universität Berlin, Germany
    https://polymake.org
@@ -41,9 +41,9 @@ class HSV;
 */
 class RGB : public GenericStruct<RGB> {
 public:
-   DeclSTRUCT( DeclFIELD(red,double)
-               DeclFIELD(green,double)
-               DeclFIELD(blue,double) );
+   DeclSTRUCT( DeclFIELD(red, double)
+               DeclFIELD(green, double)
+               DeclFIELD(blue, double) );
 protected:
    void verify() const;
    void scale_and_verify();
@@ -55,8 +55,8 @@ public:
    /// constructor from individual color channels
    RGB(double r, double g, double b) : red(r), green(g), blue(b) { verify(); }
 
-   /// ... variant for int parameters
-   RGB(int r, int g, int b) : red(r), green(g), blue(b) { scale_and_verify(); }
+   /// ... variant for integer parameters
+   RGB(Int r, Int g, Int b) : red(double(r)), green(double(g)), blue(double(b)) { scale_and_verify(); }
 
    /// converts from HSV representation
    RGB(const HSV&);
@@ -86,9 +86,9 @@ struct redirect_object_traits<RGB>
 */
 class HSV : public GenericStruct<HSV> {
 public:
-   DeclSTRUCT( DeclFIELD(hue,double)
-               DeclFIELD(saturation,double)
-               DeclFIELD(value,double) );
+   DeclSTRUCT( DeclFIELD(hue, double)
+               DeclFIELD(saturation, double)
+               DeclFIELD(value, double) );
 protected:
    void verify() const;
 public:
